feature(tasks): Move supabase logic to service and polish created-by tab usage
