+++ src/enums/task-list-view.enum.ts
@@ -1,0 +1,6 @@
+export enum TaskListView {
+    All = 'all',
+    Active = 'active',
+    Completed = 'completed',
+    CreatedBy = 'created by',
+}
