+++ src/router/index.tsx
@@ -1,4 +1,4 @@
-import TaskView from '@components/tasks/TaskView';
+import TaskView from '@components/tasks/task-view/TaskView';
 import Root from '@layouts/Root';
 import About from '@pages/About';
 import Dashboard from '@pages/Dashboard';
